Add tests for new views
