+++ tests/study_group/test_study_group.py
@@ -1,24 +1,8 @@
 import pytest
-from study_group.models import StudyGroup, GroupMember
+from study_group.models import GroupMember
 from django.contrib.auth.models import User
 from django.db.utils import IntegrityError
 from django.core.exceptions import ValidationError
-
-
-@pytest.fixture
-def make_study_group_of_varied_size():
-    def _make_study_group_of_varied_size(group_members_num, max_capacity):
-        new_group = StudyGroup(
-            group_owner=User.objects.create(username='group_owner', password='pass'),
-            field="field",
-            group_description="description",
-            capacity=max_capacity)
-        new_group.save()
-        for i in range(group_members_num):
-            new_group.join_group((User.objects.create(username=f'user{i}', password='pass')))
-        return new_group
-
-    return _make_study_group_of_varied_size
 
 
 @pytest.mark.django_db
